Move UserState stuff into subfolder
